+++ core/src/main/java/org/apache/mahout/math/hadoop/similarity/vector/DistributedPearsonCorrelationVectorSimilarity.java
@@ -26,7 +26,7 @@ public class DistributedPearsonCorrelationVectorSimilarity extends AbstractDistr
 
   @Override
   protected double doComputeResult(int rowA, int rowB, Iterable<Cooccurrence> cooccurrences, double weightOfVectorA,
-      double weightOfVectorB, int numberOfColumns) {
+      double weightOfVectorB, long numberOfColumns) {
 
     int count = 0;
     double sumX = 0.0;
